github(bin, workflows): bump GCC from 9.4 to 10.3 (#591)
The Linux jobs currently use `ubuntu-20.04`, which uses GCC 9.4 by
default. However, GCC 10.3 is also already installed [1], and with this
commit we switch to it via `update-alternatives` [2]. It is not
sufficient to only set the `CC` environment variable.

This commit does not change the Linux binary size, or produce a notable
performance improvement. But the upgrade may help for future code, or if
we move to `--mm:orc`, or start using profile-guided optimization (PGO).
See also the GCC 10 release notes [3].

At the time of writing, the `ubuntu-22.04` label is available as a
public beta [4], and it uses GCC 11.2 by default [5][6]. The latest
release is GCC 12.1.

Since GCC 5.1 (2015), GCC receives a major version bump once per year
(in late April or early May). Before that, "the GCC major number carried
little to no useful information" [7].

Here is an abridged history of recent GCC releases [8]:

    Version  Date
    -------  ----------
       12.1  2022-05-06
       11.3  2022-04-21
       11.2  2021-07-28
        9.4  2021-06-01
       11.1  2021-04-27
       10.3  2021-04-08
       10.2  2020-07-23
       10.1  2020-05-07
        9.3  2020-03-12
        9.2  2019-08-12
        9.1  2019-05-03
        8.1  2018-05-02
        7.1  2017-05-02
        6.1  2016-04-27
        5.1  2015-04-22

With the configlet 4.0.0-beta.2 Linux release binary:

    $ readelf -p .comment configlet

    String dump of section '.comment':
      [     0]  GCC: (Ubuntu 9.2.1-12ubuntu1) 9.2.1 20191022
      [    2d]  GCC: (Ubuntu 9.4.0-1ubuntu1~20.04.1) 9.4.0

With a configlet built in the GitHub Actions environment using the
changes in this commit:

    $ readelf -p .comment configlet

    String dump of section '.comment':
      [     0]  GCC: (Ubuntu 9.2.1-12ubuntu1) 9.2.1 20191022
      [    2d]  GCC: (Ubuntu 10.3.0-1ubuntu1~20.04) 10.3.0

[1] https://github.com/actions/virtual-environments/blob/984b0a83de5b/images/linux/Ubuntu2004-Readme.md
[2] https://manpages.ubuntu.com/manpages/focal/en/man1/update-alternatives.1.html
[3] https://gcc.gnu.org/gcc-10/changes.html
[4] https://github.blog/changelog/2022-05-10-github-actions-beta-of-ubuntu-22-04-for-github-hosted-runners-is-now-available
[5] https://packages.ubuntu.com/jammy/gcc
[6] https://github.com/actions/virtual-environments/blob/984b0a83de5b/images/linux/Ubuntu2204-Readme.md
[7] https://gcc.gnu.org/develop.html
[8] https://gcc.gnu.org/releases.html

Closes: #193