generate: don't alter `#` line inside fenced code block
Previously, `configlet generate` assumed that every line beginning
with a '#' character inside a concept `introduction.md` is a header.
For our purposes, the most common exception to that idea is probably
a comment inside a fenced code block. With this commit, we no longer
add a '#' to such a line.

The markdown handling here is rudimentary, but it may be good enough to
avoid a full markdown parser for `configlet generate`.

But this doesn't support other blocks, like HTML blocks.