nimble: refactor to use a `git apply` helper
Make it easier to patch other files in the future.

The new proc deliberately avoids taking a package name as an argument,
so that it supports patching a path that isn't in an installed package.
We need that for patching libcmark, because nimble doesn't copy
submodules into the installed package directory.