tests: binary: completions: fix test
The previous commit adds a test that works locally, but not in the
GitHub Actions environment. There, the start of the output (`Error`) is
printed in red. So `startsWith` would have to use color codes.

Let's just use `contains`, which is what we do elsewhere in this file.
We already check the exit code.