completion: add zsh (#653)
Add a zsh completion script and the ability to run

    configlet completion --shell zsh

to write that script to stdout.

This commit also bumps the version of the shellcheck action, which is
necessary to prevent the action from running `shellcheck` on the new zsh
script. Shellcheck does not support zsh, so CI would indicate an error
otherwise.

The bump must be to the latest commit - we already used the latest
release (1.1.0). An upstream commit [1] that claims to exclude zsh
scripts seems to be insufficient - we need to use the `ignore_names`
input that does not exist in the latest release [2].

[1] https://github.com/ludeeus/action-shellcheck/commit/ceeca77f6538
[2] https://github.com/ludeeus/action-shellcheck/commit/c2b45ddc5f85

Closes: #633