github: set draft release title and notes
Before this commit, the draft release was created without a title or
release notes. To give a better starting point for a maintainer
preparing a release, set the title to the current tag, and set the
initial notes to a list of commits since the previous official release.

The checkout Action only fetches one commit by default, so also change
the build workflow to fetch all git history.

The `gh release create` command does have a `--generate-notes` option,
but the implementation in this commit gives full control over the
format, and includes commit refs.