tests(all_tests): find via `walkDir`, not `git ls-files`
This is somewhat simpler, but may become trickier if we add a
subdirectory that contains test files (we'd need to add some
`if kind == pcDir` condition).

One advantage of the `git ls-files` approach was that it automatically
avoided finding test files in subdirectories for track repos that are
created when running the integration tests.