helpers, tests(all_tests): refactor via scanf
This is a more flexible design. We specify the pattern at the call site,
rather than hard coding it, so we can use the proc elsewhere.

As a follow-up, I plan to refactor this `getSortedFiles` and the
existing `getSortedSubdirs` to use a common implementation, where the
latter passes `pcDir` rather than `pcFile`.