github: set draft release title and initial notes (#667)
Before this commit, the draft release was created without a title or
release notes. To give a better starting point for a maintainer
preparing a release, set the title to the current tag, and set the
initial notes to a list of commits since the previous official release.

The checkout Action only fetches one commit by default, so also change
the build workflow to fetch all git history.

It seems like it's best to use the GitHub GraphQL API to get the sha of
the commit corresponding to the latest official configlet release.

With the REST API it seems like it requires two requests:

    previous_release="$(gh api repos/exercism/configlet/releases/latest --jq '.tag_name')"
    previous_release_sha="$(gh api "repos/exercism/configlet/commits/${previous_release}" --jq '.sha')"

which is careful to get the sha corresponding to the commit itself,
rather than the annotated tag (which is an object with its own sha).