patched_stdlib(json): bump from c17baae to b6bfe38 (#668)
The development version of Nim was no longer able to build configlet,
producing this error:

    $ nimble build
    /foo/configlet/src/patched_stdlib/json.nim(237, 3) Error: undeclared identifier: 'shallowCopy'

To resolve the problem, this commit pulls the latest changes to
`json.nim`. The same diff is produced when running the following
command in the Nim compiler repo:

    git diff c17baaefbcff..b6bfe38ff528 -U0 -- lib/pure/json.nim

Note that `parsejson.nim` was already up to date.