config: disable BareExcept warnings
The bump to Nim 1.6.12 caused [1] warnings to appear in the output of
`nimble build` and `nimble test`, due to the usage of bare `except` in
pkg/jsony and std/unittest.

These warnings will be disabled by default in the next Nim release [2],
and jsony may resolve the underlying issue [3], but let's disable them
explicitly in the meantime.

[1] https://github.com/nim-lang/Nim/commit/f01ffbf6f1f6
[2] https://github.com/nim-lang/Nim/commit/69c193e5e2db
[3] https://www.github.com/treeform/jsony/pull/58