lint(track_config): change `average_run_time` from a float to an int
Having the average run time as a float gives the impression of being
exact, whereas the actual run time wildly varies due to a wide variety
of reasons.

A second downside is that `jq`, which is often used to work with
track config files, will remove any trailing `.0` from a number,
which caused linting to fail.