uuid: avoid object
Arguable readability improvement: remove the need to write `.bytes`. If
the object was exported, it's not obvious that at a callsite there's
only one field.

There's no performance difference. A Nim object has no bloat - it was
still 16 bytes before.

This could be a distinct array, but borrowing the `urandom` proc and
slice operator is verbose.