empty commit
I tested the previous commit by tagging it, and pushing that tag to
exercism/configlet, which triggers the build workflow.

But it looks like it also triggered the fetch-configlet workflow, which
seems to fail on Windows [1]. Create an empty commit so that there's no
failing fetch-configlet check on the pull request.

[1] https://github.com/exercism/configlet/issues/786