lint(track_config): allow omitting concepts and exercises.concept (#832)
A recent commit [1] caused `configlet fmt -uy` to remove these keys from
the track-level `config.json` file when the corresponding value was
empty. This would cause a subsequent `configlet lint` to indicate an
error, in line with the spec. But we'll change the spec so that omitting
these keys is allowed, because:

- It's fine for tracks to lack concepts, or to lack concept exercises

- There is little value in keeping empty arrays around

- In practice the keys _are_ optional, because Exercism gracefully
  handles them being omitted 

Fixes: #831

[1] 66914845a323, 2023-10-05, "fmt: support formatting track config"