.github: delegate Zig installation to mlugg/setup-nim action (#913)
This repo uses Zig to cross-compile configlet releases, and before this
commit installed Zig via a custom install-zig script.

That script was added two years ago, back when there wasn't a great
GitHub Action for installing Zig. Since then, we've got an action by a
member of the Zig core team. Migrate to that action for installing Zig
itself, but keep a small part of the previous script for installing
zigcc.

See the mlugg/setup-zig action [1], and the 2025-06-30 Zig News post
"Community Mirrors: A Better Way To Download Zig" [2]

Closes: #912

[1] https://github.com/mlugg/setup-zig
[2] https://ziglang.org/news/announcing-community-mirrors/