config: remove now-default nimStrictDelete (#915)
The nimStrictDelete behavior is the default since Nim 2.2.0, and we've
recently updated to Nim 2.2.4 [1].

The release notes for Nim 2.2.0 [2] say:

    ## Changes affecting backward compatibility

    - `-d:nimStrictDelete` becomes the default. An index error is
      produced when the index passed to `system.delete` is out of
      bounds. Use `-d:nimAuditDelete` to mimic the old behavior for
      backward compatibility.

The configlet codebase never used `system.delete`. But Nim advertised
the upcoming behavior change, so I added `nimStrictDelete` to the
config to ensure we didn't start relying on the surprising behavior.

[1] 3e4d5c1a1968, 2025-07-28, ".github, nimble, patched_stdlib: bump Nim from 2.0.8 to 2.2.4"
[2] https://github.com/nim-lang/Nim/blob/9b527a51b854/changelogs/changelog_2_2_0.md
[3] becfece03e1d, 2021-10-14, "build(config): add -d:nimStrictDelete"